refactor: update package
- rename controller to api
- remove domain/service package
