+++ src/main/java/com/depromeet/archive/api/dto/archive/ArchiveDto.java
@@ -1,4 +1,4 @@
-package com.depromeet.archive.controller.dto.archive;
+package com.depromeet.archive.api.dto.archive;
 
 import com.depromeet.archive.domain.archive.entity.Archive;
 import com.depromeet.archive.domain.archive.entity.Emotion;
